requirements.txt
flask
google-generativeai
git+https://github.com/line/line-bot-sdk-python.git